Allow setting a name for the class report.
Besides being a nice to have, we need to be able to set a name for the class
report so that we can write to /tmp/ directory when run in containers.

Ultimately the name that matters is what's in in the S3 key which is based on
the cluster name, report month, and a timestamp.