Refactor how SU definitions are passed on
Earlier get_service_unit had a matrix with all SU definitions. In this commit
I've made it an argument that a caller should pass to get_service_unit.

This is similar to how we pass ignore_hours, so the callers like
write_metrics_by_namespace and write_metrics_by_pod can pass the su definitions.