Specify resource instead of unit in resource requests
This mainly affects the memory request because ephermeral-storage also uses
bytes as units. So, the old query would get both: memory (RAM) and ephemeral
storage requests.

Luckily, when we are merging the metrics we would always overwrite the
ephemeral-storage bytes (for pods that requested ephermeral storage) with the
memory bytes just because "e" comes before "m".

So, while we are not affected it's just best to be specific in the metrics
we want so for that I reason I also changed the CPU metrics to specify cpu
instead of the unit "cores".