Added `Cluster Name` column to invoice
Cluster name is determined by the Prometheus URL used to query the metrics.
Known Prometheus URLs are mapped to cluster names. If an unknown URL is
found, it is used as the cluster name instead.

This change supercedes the use of `OPENSHIFT_CLUSTER_NAME` [1].
All references to it has been removed.

[1] https://github.com/CCI-MOC/openshift-usage-scripts/commit/79ad4a1e8cd53c9ae36de24f946a771fa0c40352