Make setting prometheus query interval minute configurable
We have some clusters where we would like to change the frequency at
which we collect metrics and generate the report based on that.

To accomplish this, we need to make 2 changes. We set an environment
variable that is passed on to the metrics collector scripts. And we
insert the collection interval into the json file.

The merge scripts will read this interval when loading these files. For
backwards compatibility with our old metrics file, we'll default to 15
minutes when processing these files.