Remove logger warning about reports spanning multiple months
The new columns will indicate the report start and end date, so it's not
strictly necessary to log a warning anymore. Futhermore, since it
modified the report_month string, it'll break the call the to
get_su_definitions down the line. I could just move that call to it up,
but this warning isn't that necessary anyway.